requirements.txt
langchain
langchain-community
langchain-pinecone
pinecone-client
python-dotenv